refactor: add CryptoSection to SupportFundPage
